Moved responseAs test to different file as it's not the same thing as the StatusTests - it's using the status endpoint as a simple e.g. of a working endpoint, not testing that it is working
