+++ src/SevenDigital.Api.Wrapper.Integration.Tests/EndpointTests/StatusTests.cs
@@ -1,18 +1,10 @@
 ﻿using System;
-using System.Xml.Serialization;
 
 using NUnit.Framework;
 using SevenDigital.Api.Schema;
 
 namespace SevenDigital.Api.Wrapper.Integration.Tests.EndpointTests
 {
-	[XmlRoot("response")]
-	public class RawStatusResponse
-	{
-		[XmlElement("serviceStatus")]
-		public Status ServiceStatus { get; set; }
-	}
-
 	[TestFixture]
 	public class StatusTests
 	{
@@ -23,16 +15,6 @@ namespace SevenDigital.Api.Wrapper.Integration.Tests.EndpointTests
 
 			Assert.That(status, Is.Not.Null);
 			Assert.That(status.ServerTime.Day, Is.EqualTo(DateTime.Now.Day));
-		}
-
-		[Test]
-		public async void TestStatusResponseAs()
-		{
-			var statusResponse = await Api<Status>.Create.ResponseAs<RawStatusResponse>();
-
-			Assert.That(statusResponse, Is.Not.Null);
-			Assert.That(statusResponse.ServiceStatus, Is.Not.Null);
-			Assert.That(statusResponse.ServiceStatus.ServerTime.Day, Is.EqualTo(DateTime.Now.Day));
 		}
 	}
 }
