+++ src/SevenDigital.Api.Wrapper.Integration.Tests/SevenDigital.Api.Wrapper.Integration.Tests.csproj
@@ -52,6 +52,7 @@
     <Compile Include="AppSettingsCredentials.cs" />
     <Compile Include="EndpointTests\BasketEndpoint\BasketCreateTests.cs" />
     <Compile Include="EndpointTests\ReleaseEndpoint\ReleaseEditorialTests.cs" />
+    <Compile Include="EndpointTests\ResponseAsTests.cs" />
     <Compile Include="EndpointTests\StaticSetupCredentialPassingTests.cs" />
     <Compile Include="EndpointTests\OAuth\OAuthRequestTokenTest.cs" />
     <Compile Include="EndpointTests\ReleaseEndpoint\ReleaseTracksTests.cs" />
